ENH: move 'add_source', 'add_quantity', to 'EVENT' class - start integrating into OOP framework.  Moved 'clean_event' and 'load_event_from_file' into 'Events.py'.
